requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
pandas>=2.1.4
datasets>=2.16.1
sentence-transformers>=2.3.1
faiss-cpu>=1.8.0
torch>=2.1.2
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.27.0
